+++ app/components/modals/modals.js
@@ -1,0 +1,4 @@
+var angular = require('angular');
+
+angular.module('modals', [])
+  .directive('modal', require('./modal-directive'));